refactor: owner 리팩토링 (#1040)
* chore: controller 클래스 분리

* chore: dto 패키지 구조 변경

* chore: model패키지 구조 변경

* refactor: 예외처리 별도 클래스로 분리

* refactor: 서비스 로직 클래스 분리

* refactor: 공통 로직 모듈화

* test: 테스트 응답 필드 변경

* chore: 이미 존재하는 이메일인지 검증하는 메소드 이름 변경

* chore: 중복된 이름의 빈 충돌 해결

---------
