+++ src/main/java/in/koreatech/koin/domain/owner/dto/sms/OwnerSendSmsRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.owner.dto;
+package in.koreatech.koin.domain.owner.dto.sms;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
@@ -10,7 +10,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record VerifySmsRequest(
+public record OwnerSendSmsRequest(
     @NotBlank(message = "휴대폰번호는 필수입니다.")
     @Schema(description = "휴대폰 번호", example = "01000000000", requiredMode = REQUIRED)
     @Pattern(regexp = "^\\d{11}$", message = "전화번호 형식이 올바르지 않습니다. 11자리 숫자로 입력해 주세요.")
