--- src/main/java/in/koreatech/koin/domain/owner/model/OwnerRegisterBySmsEvent.java
@@ -1,7 +1,0 @@
-package in.koreatech.koin.domain.owner.model;
-
-public record OwnerRegisterBySmsEvent(
-    Owner owner
-) {
-
-}
