--- src/main/java/in/koreatech/koin/domain/owner/model/OwnerRegisterEvent.java
@@ -1,7 +1,0 @@
-package in.koreatech.koin.domain.owner.model;
-
-public record OwnerRegisterEvent(
-    Owner owner
-) {
-
-}
