+++ src/main/java/in/koreatech/koin/domain/owner/model/dto/OwnerEmailRequestEvent.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.owner.model;
+package in.koreatech.koin.domain.owner.model.dto;
 
 public record OwnerEmailRequestEvent(
     String email
