+++ src/main/java/in/koreatech/koin/domain/owner/model/dto/OwnerRegisterEvent.java
@@ -1,0 +1,9 @@
+package in.koreatech.koin.domain.owner.model.dto;
+
+public record OwnerRegisterEvent(
+        String ownerName,
+        String account,
+        Integer ownerId
+) {
+
+}
