+++ src/main/java/in/koreatech/koin/domain/ownershop/service/OwnerShopUtilService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class OwnerUtilService {
+public class OwnerShopUtilService {
 
     private final ShopRepository shopRepository;
