+++ src/main/java/in/koreatech/koin/domain/user/repository/UserRepository.java
@@ -27,7 +27,7 @@ public interface UserRepository extends Repository<User, Integer> {
 
     default User getByEmail(String email) {
         return findByEmail(email)
-            .orElseThrow(() -> UserNotFoundException.withDetail("email: " + email));
+            .orElseThrow(() -> UserNotFoundException.withDetail("account: " + email));
     }
 
     default User getByPhoneNumber(String phoneNumber, UserType userType) {
