+++ src/main/resources/db/migration/V18__alter_user_email_column_nullable.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `koin`.`users`
-    CHANGE COLUMN `email` `email` VARCHAR(100) CHARACTER SET 'utf8mb3' NULL COMMENT '학교 email' ;
+    CHANGE COLUMN `email` `email` VARCHAR(100) CHARACTER SET 'utf8mb3' NULL COMMENT '학교 account' ;
