+++ src/test/java/in/koreatech/koin/acceptance/UserApiTest.java
@@ -393,7 +393,7 @@ class UserApiTest extends AcceptanceTest {
 
     @Test
     void 이메일이_중복인지_확인한다_잘못된_이메일_형식이면_400() throws Exception {
-        String email = "wrong email format";
+        String email = "wrong account format";
 
         mockMvc.perform(
                 get("/user/check/email")
@@ -592,7 +592,7 @@ class UserApiTest extends AcceptanceTest {
                     .content("""
                         {
                           "major": "컴퓨터공학부",
-                          "email": null,
+                          "account": null,
                           "name": "김철수",
                           "password": "cd06f8c2b0dd065faf6ef910c7f15934363df71c33740fd245590665286ed268",
                           "nickname": "koko",
@@ -614,7 +614,7 @@ class UserApiTest extends AcceptanceTest {
                     .content("""
                         {
                           "major": "컴퓨터공학부",
-                          "email": "koko123@gmail.com",
+                          "account": "koko123@gmail.com",
                           "name": "김철수",
                           "password": "cd06f8c2b0dd065faf6ef910c7f15934363df71c33740fd245590665286ed268",
                           "nickname": "koko",
@@ -636,7 +636,7 @@ class UserApiTest extends AcceptanceTest {
                     .content("""
                         {
                           "major": "컴퓨터공학부",
-                          "email": "koko123@gmail.com",
+                          "account": "koko123@gmail.com",
                           "name": "김철수",
                           "password": "cd06f8c2b0dd065faf6ef910c7f15934363df71c33740fd245590665286ed268",
                           "nickname": "koko",
@@ -655,7 +655,7 @@ class UserApiTest extends AcceptanceTest {
                     .content("""
                         {
                           "major": "컴퓨터공학부",
-                          "email": "koko123@gmail.com",
+                          "account": "koko123@gmail.com",
                           "name": "김철수",
                           "password": "cd06f8c2b0dd065faf6ef910c7f15934363df71c33740fd245590665286ed268",
                           "nickname": "koko",
@@ -740,7 +740,7 @@ class UserApiTest extends AcceptanceTest {
                 post("/user/login")
                     .content("""
                         {
-                          "email": "%s",
+                          "account": "%s",
                           "password": "%s"
                         }
                         """.formatted(student.getUser().getEmail(), "1234"))
